refactor: isolated hook's helper function
